Add support for standard json serialization
The company search by domain turns out to be a simple Json serialization
of the incoming DTO with no need for all the black magic in our custom
serializer.

So, allow calling the serializer with just an object, in case it is an
instance of a IHubSpot* we throw it through the custom serializer, if not
we just use the JsonConvert directly.