Fix malformed URL when listing contacts
The URL was malformed and the underlying http request builder did
not take into account that the various clients could inject base urls
that already contained query params.

Pulled in flurl to allow dynamic building of query parameters on the
url that we are firing towards HubSpot.

Added integration test to show that list contact now works as intended