Update ListOfContactListsHubSpotEntity.cs
Change DateTime properties to string because deserialization of .NET type DateTime in WCF requires some more work e.g. the use of attributes [IgnoreDataMember], [OnSerializing] and [OnDeserialized] along with a private property - see https://stackoverflow.com/questions/18511147/change-default-date-serialization-in-wcf. String datestamps can be easily translated to DateTime explicitly by the caller as necessary. Adding DataMember attributes revealed the problem of deserialization of the DateTime properties (but not before runtime) - it caused exceptions where as without the DataMember attribute, the deserialization failure was not notified.